implement toy eager and lazy file systems, and try to prove them correct.
the eager file system is proven correct according to the spec.

what previously seemed to be a weirdness in my new spec turns out to
be a critical problem that prevents proving the lazy file system correct.

in particular, the spec in effect requires that Sync force the last
Write to disk, even if that Write happened before a Crash.  but the
lazy file system forgets the in-memory state after a Crash, and since
it forgot the last (pre-Crash) Write, it cannot force it to disk..