Start a better version of fs_apply_list
applyfun used to manipulate the trace directly. It could provide
fs_legal with a valid trace even though it does not function correctly.

The new definition of applyfun acts as a monadic state transformer:
    applyfun :  fsstate -> invocation -> fsstate * result
and fs_apply_list extends the trace based on the invocation and result.

fs_apply_list should also validate the returned result is a valid
response for the invocation.  The next step is to eliminate trace
from the return value altogether: fs_apply_list returns a final state
and a proposition proving that the trace is legal.