Adam-style transactional disk
The high-level language tprog operates on two hypothetical disks:
one for normal read/writes (nx-disk), the other (alt-disk) for
read/writes inside a transaction.

TBegin copies nx-disk to alt-disk; writes inside a transaction
are made on the alt-disk (i.e. isolation); TEnd replace nx-disk
with the content of alt-disk (i.e. atomicity). The correctness
of this level is easy to verify --- if tprog halts inside a
transaction, the nx-disk is left intact.

This abstraction can be easily extended to support multiple
transactions, by introducing multiple alt-disks organized as
a stack (for nested tx) or a map (for concurrent tx).

The low-level language pprog implement the alt-disk with an
in-memory pending list and a tx flag. Primitives to access them
include GetLog, GetTx, AddLog, ClrLog and SetTx. We choose these
fine-grained operations so that during further refinement, they
can be implmented in the lower level with real disks efficiently.

pprog level could crash, for now it halts after running the
recovery procedure (which we assume does not fail).
In real world, we want the system to continue running after recovery,
and the recovery procedure could fail too. Maybe we can model this
by simply changing the stepping rule (PsCrash)?