BufferedFileSpec: sketch a non-deterministic buffered file spec
there's two parts to the state: the on-disk state and the in-memory state.

all operations except Sync behave identically to the synchronouse FileSpec;
in fact, they inherit the inductive predicates from FileSpec.

the Sync operation succeeds only if the on-disk state is already the same
as the in-memory state.

there's also a "flush" step, which does not consume any program code, but
instead copies the state of a single inode from ephemeral state to persistent
state.

since the "flush" step does not consume any program, it can happen any time,
and is the source of non-determinism in this specification.