slight change to the ";" and ";;" notations
Use a dummy definition to "remember" when something was the result
of using the ";" or ";;" notations, so that Coq knows to print them
back correctly (and in particular, not display all function applications
using the ";;" notation).  This seems to have virtually no effect on
anything else, including proofs; Coq seems smart enough to unfold the
dummy definitions as needed.