use proof irrelevance to prove star_do_iwrite_blocklist
If we used Z instead of nat, we could get away with not relying on
proof irrelevance, because Z.lt is defined as equality of Z.compare
to Lt, and then you can apply eq_proofs_unicity.  Unfortunately,
nat's lt is not done this way, so it seems like we have to invoke
proof irrelevance.