checkpoint some experiments with hoare logic
get the log system to have a more familiar interface (basically,
introduce the extra on-disk variable that adam suggested).

converted most of the spec, impl, and proofs to this new interface.

stuck at a weird hoare logic issue in the recover_ok proof: it
appears that the hoare tactic somehow ends up matching the ghost
variable in the apply loop invariant to m' rather than m, which
should be impossible: apply is invoked only if LogCommit=1, and
that's possible only if the (rep xp (CommittedTxn m)) part of
the precondition holds.  but maybe this is some misunderstanding
on my part..

a more mundane issue: how to invoke programs from other programs?
currently, recover and commit both invoke apply; to make this work,
the code explicitly constructs apply as a prog' rather than using
the $() syntax for a prog.