Checkpoint an attempt at commit'ing a non-deterministic state.
Naive approach to specify non-determinism: in the post-condition of the
function (which we're about to wrap in a transaction), assert that the
resulting mem is just some mem that satisfies some externally-supplied
predicate.

This approach fails, for a somewhat surprising reason:

The post-condition in question gets copied in several places during
the proof (perhaps by spost, perhaps by the hoare tactic, perhaps by
something else).  There's nothing that requires the different copies
of this post-condition to produce the same mem; only that all of them
satisfy the externally-supplied predicate.  As a result, the proof gets
stuck with two possibly-different mem's satisfying the same predicate.