Getting rid of prog'
Instead, try to prove theorems (corr statements) about each higher-level
construct like If and For, as Adam suggests.

Proved theorems about If and Call; now stuck on For.

Also proved a helper theorem:

  forall pre pre' pr rec,
  (pre ==> [{{pre'}} pr >> rec]) ->
  (pre ==> pre') ->
  {{pre}} pr >> rec.

which was useful for Call, but unfortunately not for If (seemingly due
to the exists in the If's precondition).