go back to using the same "mem" type inside the txn as on the outside
this means we might be able to use |-> for in-transaction memory
(e.g., in transactional read/write).

not quite clear what's the best way to do this sort of "nested separation
logic"; see the awkward-looking precondition in LOG.read_ok/LOG.write_ok.

almost certainly the automation can't deal with this nesting either, yet.