checkpoint a better rep invariant for the log
represent the log as a list of log entries.  the list always has a
fixed number of log entries (LogLen xp).  only a prefix of the log
entries is considered for ActiveTxn.  the list of log entries maps
to a bunch of sep_star'ed ptsto pred's.

still need to figure out a way to access an individual ptsto pred,
so that we can automate reasoning about writes to log locations.