checkpoint some progress on log invariants
changed the log's representation invariant from a list to a map
(nat->logentry).

introduced theorems to split out a particular log entry from the folded
sep_star's, and to merge it back in.

used these theorems in write_ok to both split out and merge back in a
given log entry.

not quite clear how to automate this yet.