bizzare ltac issue
in proving apply_ok, we have:

  Ltac hhh := hint_avail_region_grow.
  Ltac norm_hint2 hhh := repeat norm_or_l; norm'l; try hhh; try norm'r.

trying to invoke "norm_hint2 hhh" or "norm_hint2 ltac:(hhh)" produces:

  In nested Ltac calls to "norm_hint2" and "norm'l", last call failed.
  Error: Tactic failure: Incorrect number of goals.

but copy-and-pasting the body of norm_hint2 produces no errors!

Adam suggested two possible workarounds:
- adding ltac:(...) around some tactics
- adding idtac; in front of some tactics
but neither seem to help (and the first one doesn't seem to parse
in the body of the Ltac definition for norm_hint2).