perf: use [trivial with okToUnify] instead of auto
speeds up proofs when there's an expensive hint for okToUnify of two
terms (e.g., it includes [ring]), the hint doesn't work out, and [auto]
keeps trying something-or-other and invoking this hint over and over.

using [trivial] instead of [auto], with all okToUnify hints now of cost
zero, means we try each hint just once, so it's not so damaging if the
hint's tactic is slow.