finally, prove log recover idempotence!
the resulting theorem is a little more hairy than expected, but at least
it's a start:

  Theorem recover_idem_ok : forall xp rx,
    {{ exists m F rec, log_intact xp m F
     * [[ forall rec',
          {{ rep xp (NoTransaction m) * F
           * [[ {{ log_intact xp m F }} rec' >> Check (log_intact xp m F) ;; rec' ]]
          }} rx tt >> Check (log_intact xp m F) ;; rec' ]]
     * [[ {{ log_intact xp m F }} rec >> Check (log_intact xp m F) ;; rec ]]
    }} recover xp rx >> recover xp rx.

the "exists rec" is a little awkward (it should just be "recover xp rx"),
but i couldn't yet prove it's unnecessary.  if we had a better plan for
proving liveness (i.e., that a particular continuation will get executed),
this might not be necessary.

there's also one more admit left in apply_ok; probably some issue with
the automation machinery, but it takes well over 10 minutes to get to
that point in the proof script, so it's quite painful to debug..