checkpoint more progress on the new plan
lots of unification problems arise.  the most worrisome seems to be
the one that shows up in [commit_ok].  the theorem statement says
something like [rep xp (CommittedTxn m2) * F ==> crash].  in the
middle of the proof, we get this as a hypothesis, and also as a goal.

unfortunately, there's an [exists] inside CommittedTxn, which gets
turned into a separate variable in the hypothesis, and a different
separate variable in the goal.  now this ends up being unprovable.