an almost-proven-correct simple inode impl
one admit remaining: proving that updating a single inode in a block of inodes
does the right thing (in iput).

this inode API relies on Pack.v to update/extract inodes from a block,
and on Rec.v to convert between words on disk and the inode structure.
one of the Pack.v theorems is still unproven due to some dependent
type mess.

this inode API also doesn't yet have unified inode numbers.  instead,
inodes are identified by a tuple of (block containing the inode, offset
within that block).  i expect we should wrap iput and iget with code that
translates between flat inode numbers and these tuples, along with a proof
that the translation is a bijection, and that the representation can be
thought of as a flat list of inodes, rather than a two-dimensional list.

next, should probably implement interesting data within an inode (i.e.,
store something interesting in the inode's one-and-only data block)..

Coq bug 3731 keeps biting me, making "rewrite_strat" take forever on
certain rewrite hints, necessitating less use of "Hint Rewrite", and
explicit avoidance of "rewrite_strat" in favor of "autorewrite with core",
but that's super slow for Log.v, so we can't switch everything back to
"autorewrite with core"..