morphisms seem promising!
this example uses [a =p=> b] for what used to be [a ==> b]

Example pimpl_sep_star_rewrite : forall a b x y, a =p=> b
  -> (x * a * y =p=> x * b * y).
Proof.
  intros.
  rewrite H.
  reflexivity.
Qed.

this might make all of our norm hints much less annoying, because we
could do a straight-up rewrite, instead of monkeying with indirect hints
to the automation machinery..