partially clean up iput_ok
autorewrite_fast now has a workaround for Coq bugs 3731/3798, so
no need to expand the first manual step in iput_ok.

still need the explicit [autorewrite with core in *] because of
Coq bug 3312.

and still need the manual unfolding of the second step, because
type checking takes forever due to some expansion of addrlen.
one cause was expansion of wlt_dec, but even with that made opaque,
other things still expand addrlen.