Fix handling of nested predicates.
This helps with proving iget_ok.  Now that proof is stuck at other
points that are more relevant to the theorem in question, like proving
that division and modulo have certain properties, and that the inode
selected out of the flat list is the same as the one selected out of
the two-dimensional list.  (Admitted all of those for now.)