Write inode_match as sep logic
Almost proved iput_ok.  The messy part is still in RecArray.
Array.data is internally represented as pairs.
In `step`, `destruct_lift` destructs the pair and generates
lots of `let` express in the goal.  It is hard to fold them back
or to write theorem about them.

Is there a way to make Rec.data opaque to hoare tactics?