checkpoint trying to prove read_ok
tricky because the disk can decide to flush some data on its own
(XSelfFlush), and that might affect the frame predicate.  so we
may need to restrict the kinds of frame predicates we allow?  or
go back to a less-faithful exec model where the disk is not allowed
to flush on its own, and does flush per sector?