More separation-logic based specs
By writing precondition like this
    (F * inum |->?)%pred (list2mem ilist),
we get the bound (inum < length ilist) for free.  This is nice.

However, inside the proof, we usually need to convert back and
forth between separation logic representation and `sel`, `upd`
style.  This is especially painful when `upd` is involved.
We need a better way to automate this.