extract [nat] to [Int] rather than [Integer]
as a very rough estimate, our performance seems to be within ~10x of what
we should expect for our design.  on linux, a vaguely similar benchmark
(grow-write-read-shrink on an existing file descriptor) takes 0.07s,
in contrast to our 1.6s, but we are doing more (creating log entries
for two transactions in each iteration, in contrast to linux updating
the file cache).  perhaps the xv6 file system would be another good
baseline for performance comparisons.