Checkpoint progress of proving indirect blocks
The problem I have now is to construct a block list from a disk
valu returning from the block allocator.  Specifically, I need
to proof:

        a |-> v  =p=> array_item _ ?blist,

where `a` is the newly allocated block, `v` can have arbitrary value,
and ?blist is an existential from `indput`'s pre condition.  I need a
way to construct ?blist from v, and prove that it meets RecArray's rep
invariant.

This is the first time we encounter this problem.  Before this,
everything we were reasoning about already "exists" on the disk:
we reserved space for inode/file, bitmap, log, etc.  But the new
indirect block list seems to come from nowhere, so there is no
"pre-existing" blist in the spec I can exploit.

Other issues:

- Write subroutines that preserve the representation invariant.

  I use to split igrow into 3 pieces that pass around an inode'
  object, which represents intermediate changes to the inode's
  state that doesn't yet appear on the disk.  This turns out
  to be a bad idea: the subroutine's spec cannot be written in
  in sep-logic, but has to peek into the intermediate states.
  This make the spec very hard to write, and hard to converge
  back to the separation-logic style.

  This encourages a programming style of writing separate
  routines for different cases.  Even though two cases can share
  most of the code, they cannot call a common subroutine that
  operates an intermediate state.  This is a big bummer when
  the program's logic gets complicated.  What's the better way to
  deal with this?

- Currently our "If" construct doesn't have a "continuation"
  argument as "For" does.  So there's no way to merge control
  flow after the branch unless we start a new subroutine.