Use partial evaluation to eliminate Rec's ASCII strings at runtime.
Use [Eval .. in ..] to define our executable programs in Inode.v,
with the [cbn] convtactic (wrapped as [compute_rec], in case we
need to change it later).

This produces much cleaner (and somewhat faster) extracted code,
because we no longer compare ASCII strings at runtime.  Calls to
recget' and recset' are instead replaced with let statements.

This might also help address Haogang's concerns about the ASCII
strings leading to very slow proofs.  Unfortunately, the current
proofs in Inode.v are probably broken as a result of this change
(sorry!), but the proofs weren't working for me before this change
either..