Provide a general version of Haogang's [rec_simpl] tactic from Inode.v
Rec.v now provides a [rec_simpl] tactic that replaces all of the let
statements generated by the new [compute_rec] plan with uses of [fst]
and [snd].  This transformation makes it easier to manipulate the
resulting terms.

This generalized [rec_simpl] tactic seems to be a close replacement for
the old inode-specific [rec_simpl].  For example, the original proof
for [isetsz_ok], which used the old [rec_simpl], now works again with
no changes.