introduce notion of complete predicates
if a complete predicate holds over two memories, those memories must
be equal.

prove that [ptsto] and [emp] are complete predicates.

prove that if you use a complete predicate in listpred, and two
memories match listpred with the same list, the two memories are
also equal.

will be useful in Dir.v shortly.