optimize the extracted code for the cache
unfortunately, even with Coq [positive] and [Z] mapping to
Haskell's Integer, the cache is still a net loss in terms
of performance: with the cache, we get ~36MB/sec, and the
last numbers before the buffer cache were ~52MB/sec.  bummer!

quite a bit of time is spent rebalancing FMapAVL trees.