Prove generalized bitmap allocator spec, good for any [addr]-indexed separation logic memory
The specification for the generalized allocator, [BALLOC.alloc_gen_ok] and
[BALLOC.free_gen_ok], says that there's a predicate associated with every
allocator state.  Their postconditions say that, when [BALLOC.free_gen] is
called, this predicate grows by one block (the freed block).  Conversely,
when [BALLOC.alloc_gen] is called, the predicate shrinks by one block
(the allocated block).

Instantiating this generalized bitmap allocator for allocating disk
blocks is easy: this predicate is simply [sep_star]'ed into the same
memory that's storing the disk blocks.  Indeed, this patch also builds
the regular block allocator in this style, with surprisingly short proofs.

Instantiating the generalized allocator for inodes should be "just"
a matter of [sep_star]'ing this predicate into the inode memory instead
of the disk block memory.

This brings up an interesting point: it seems like the inode allocator
should be part of the directory layer -- not the inode layer!  From this
generalized allocator point of view, allocation is a question of who
owns the inode: either the allocator (through this shrinking/growing
predicate), or something else (through some other predicate that should
probably look like BFile's [file_match] and [data_match]), and this
something else is probably a directory entry.

It seems like we should write a predicate saying that each filename maps
to a unique inum in the inode memory.  Supporting hard links seems hard
because then multiple directory entries own a single inode, and that's
tricky to represent in separation logic.