Prove bf_getlen_ok.
Also simplify preconditions for the index argument in
bf_put and bf_get.

Lots of effort to show that the number of items in BFileRec does
not overflow. i.e.

    items_per_valu * INODE.blocks_per_inode < pow2 addrlen

To be general, we prove that

    valulen * INODE.blocks_per_inode < pow2 addrlen

The problem is when a constant is bigger than 2^15, `compute`
will cause stack overflow.  Our current valulen is already at the
boundary.  So I have to prove bounds for both multipliers and then
combine them together.

Not sure what's the better way of handing large constant value in Coq.