Dir, DirName: try to avoid such giant [rep] functions..
Dir's [rep] used to include all of [BFILE.rep], which made it difficult
to reason about multiple Dir's in DirTree.  this fix tries to disentangle
the important parts of Dir's [rep] from the convenience macro that includes
the entire stack from [BFile] down to [Dir], which is now called [rep_macro].
[DirTree] will use only [rep], because it has its own plan for managing the
entire [BFILE.rep].

the same fix applies to DirName as well.