hslib/Word.hs: fix a subtle [split1] corruption bug in haskell
my haskell impl of [split1] was assuming that when the word was being
split into two pieces (sz1 and sz2 bits respectively), the source word
was already taking up at least (sz1/8) bytes of memory.  this was not
true for words that have lots of leading zeroes!  consequently, this
led to all kinds of random garbage being introduced into these words,
because we were over-reading memory..