Dir.v: complete all proofs
Change the post condition of [deext_ok] to explicitly export the new [delist].

Since we use [listpred dmatch] to construct the new dmap in [dlink_ok],
it's more convenient to have the explicit form of the new [delist] in hand.
Generalized separation logic (list2nmem) turns out to be useless here.