Cache.v: clean up proofs and implement validator-based replacement algorithm
The cache now invokes an external replacement algorithm to keep track
of what blocks are being accessed, and to choose a block when evicting.
If the replacement algorithm malfunctions, and returns a block that is not
in the cache, the cache code will choose its own block to evict instead.

Amusingly, because of Haskell's lazy evaluation, even though every
disk access is supposed to update the replacement algorithm state, it
never gets evaluated (because we never evict in our workloads), so the
fact that there is no actual implementation doesn't bother Haskell --
the file system continues to work!  I expect it will throw an exception
when it needs to actually evict some block, and then we can implement
some replacement algorithm.

Also add a paragraph to the paper about this validating replacement
approach.

A few admits left about lengths of [Map.cardinal].  I decided to use [nat]
instead of [word] for keeping track of the current cache size.  Since this
value is not exported to/from disk, there's no reason to keep it as
[word].  Avoiding [word] made the proofs much easier and much faster.