DirName.v: prove [dslookup_ok]
[pred_apply] sometimes generalizes predicates too aggressively
and might lose some information.  For example:

    m', m: mem
    name: string
    H1: forall x, indomain x m -> name_valid x
    H2: name_valid name -> False
    H2: mem_trans_a f m' m
    ......
    ---------------
    notindomain name m

The goal only holds for a particular memory [m], given the information
inside H1.  H2 is totally irrelevant here.  However, [pred_apply] will
turn the goal into an unsolvable one:

    ---------
    mem_trans_a f m' =p=> notindomain name

Current solution is to carefully step through, and use [cancel']
instead of [cancel] when this might happen.  Is there a better way
to handle this?