Almost proved dslink_ok
Hit limitations of separation logic again when proving [dslink_ok]
and [dsunlink_ok].  To prove them, we have to explicitly expose the
updated memory (dmap') in [dlink_ok] and [dunlink_ok] in Dir.v.

The problem is, [mem_atrans] operates on the entire [dsmap] memory,
not just on a single directory entry.  To prove its property, we have
to know exactly what the entire memory looks like.  However, separation
logic frame [DF] is too weak to infer this information from.

The same problem happens in Dir.v as well, when we try to construct the
directory map memory from a list of directory entries.  Whenever we
update a single dirent, we have to reconstruct dmap using the entire
list.  So we have to know the exact form of the updated dirent list.
Anyway, sep-logic is good at isolating elements of interest, rather than
reasoning about the entire memory.

To keep our spec clean, we might want to introduce the notion of
"complete predicates F", such that F m -> F m' -> m = m'.
Then we can prove some theorems for backward reasoning, such as:
    (F * a |-> v) m  ->  F m'  ->  m' = mem_except m a
    F m  ->  (F * a |-> v) m'  ->  m' = upd m a v