flesh out [mkfile_ok] and [mkdir_ok] slightly further
looking actually pretty easy!  the remaining parts are to actually
initialize the new file/subdirectory to zero length (we weren't doing
this before, so it's cool the proof caught it), and to prove that a
zero-length new subdirectory is empty.

the other step will be to prove that, once we update the local tree
representation, we can splice it back into the complete tree.  that
seems harder.