Found one reason why [step] is slow
In SepAuto.v we call [ring] to decide okToUnify for [ptsto],
this can be very slow sometimes.  It might even get stuck for
this pattern:
   a |-> v =p=> # $ a |-> v

Add a timeout for [ring] since I'm not sure whether other proofs
rely on it.  However, it's better to keep okToUnify lightweight
or at least make the hint local.