introduce [would_recover_either_pred]
ran into an unfortunate problem in proving statements like

  exists m', would_recover_either m m'

because in some situations, there might not exist such an m';
in that case, of course we wouldn't crash in that m' state either,
but the statement is unprovable.

instead, the plan is to write things like:

  would_recover_either_pred m (DIRTREE.rep xx * [[ .. ]])

where the last argument applies to m', in cases when we're not in m.