[FS.mkfs]: comment out call to [INODE.init]
unfortunately, our inode size has grown, and as a result, writing
zeroes to all the inode blocks overflows the max size of a transaction
(limited by a single header block).  for now, let's just rely on the
"dd if=/dev/zero" in our benchmark script to do the same thing.  later
we can fix this properly in MemLog..