try to fix the theorem statement for [flush_sync_ok]
the new version should be provable, and the troublesome case does indeed
prove now, but i haven't tried to go through the rest of the proof steps
(they should work, but might be broken for mechanical reasons).

this change might also impact [flush_ok] in a small way; we should check
that it still proves afterwards..  but [flush_ok]'s crash condition already
includes everything in the new [flush_sync_ok] crash condition, so it should
still be provable.