fix Go extraction of mutually-recursive fixpoints
Go prohibits loops in global variable initialization, so when we define
several mutually-recursive fixpoints, just duplicate the code for the
other functions when defining each function in turn.

all Go code for FSCQ actually compiles now!  need to write some kind of
driver now, and see if we need to optimize the handling of closures.