Add _CoqProject to pass arguments to Coq
These arguments are used by Coq IDE if the setting is enabled under
Project (the default is to ignore the file). My setup uses the
pre-packaged 8.5 beta 2 app, which doesn't work when launched with the
desired arguments for Coq, but this fixes that problem.

If this causes you problems, let me know and I'll remove and gitignore
this file. Otherwise, it could obviate the need to use coqide.sh to
launch CoqIDE.