Split bf_resize_ok into shrink/expand ok
Each has an idiosyncratic way of stating what is preserved in
separation logic. We have yet to see what higher levels will need from
these specs, but it seems reasonable to have one spec for each of the
two modes of operation and have the application choose which one to use
(eg, by having bf_shrink and bf_expand be aliases to bf_resize with
different automation hints).