Add proofs to chunks
chunkList satisfies these obligations via Program Fixpoint, which I
found more convenient than the explicit proofs provided in ByteFile's
chunks.

The proof term boff <= (block size) is redundant by transitivity so I
removed it and replaced it with an outside definition (so functionally
the same proofs are available, but the obligation burden is smaller).