Formalize write spec with preserved preds
Now says appropriate parts before and after write are preserved, and
postcondition specifies a hole for the zeros. Any of these parts could
be zero-length (eg, the write is an append, so no part of the file
after the right is preserved), which become vacuously true by having
the predicates be over zero-length lists.