Modulo some admits, finish proving bf_put_chunk_ok
The admits are in lemmas that convert update_chunk (which takes a word) to
update_chunk_block (which takes a block = a type of list). The former
uses isplit1, isplit2, and icombine while the latter uses firstn, skipn
and app.

These functions are isomorphic in the following way:
* isplit1 ~ firstn
* isplit2 ~ skipn
* icombine ~ app (= ++)
* [items count] ~ [l: list item /\ length l = count],

but proving these relationships requires some dependent types magic I
don't fully understand.