Expose some problems with apply_build_chunks
The proof strategy should probably be to first show apply_chunk is
equivalent to a bunch of updN's, then relate that to the lists (or even
migrate everything to updN, but that seems unecessary).

Manipulating firstn/skipn nicely requires length bounds that seem to be
problematic for the induction, so it's hard to prove something general
about build_chunks outside of the specific parameters used in chunkList.