Fix a bug in num_chunks computation for chunkList
Naturally, this breaks a bunch of proofs, including the big bound proof
(which now has to deal with divup).

It turned out the last chunk was being excluded and the bound took this
into account, which made proving that blocknums were in bound impossible
for small writes. Eventually we would also have found that not all the
bytes were being updated in the apply_chunks proof.