Remove duplicate combine_app proof
If we're keeping both versions of the theorem, then one proof should
use the other. The difference here is that combine_app provides an
explicit proof that the sizes work via the len_add function, which
means that it cannot be used to rewrite <- with a different proof
(despite the axiom in ProofIrrelevance we're admitting).