Modulo mostly true admits, finish grow_file_ok
The spec/implementation are problematic: we need to handle the case
where the file is already long enough, where currently the
implementation does thing and returns true.

One way to do this is introduce an existental newlen' in the
postcondition and say newlen' >= newlen, or even precisely that newlen'
= max newlen #curlen. It's unclear whether grow_file should even be
doing this check, or if the caller should deal with it.