Vastly speed up step over update_bytes
Apparently the bottleneck is actually apply update_bytes_ok (not the
rest of the search, or even eauto with prog). Manually specifying
several of the arguments to the lemma (it seems only the number of
unspecified arguments is important, not which ones) vastly speeds up
this step.

Unfortunately I haven't found a way to make this robust. It looks like
explicitly adding [mscs] screws up the hint since mscs often appears as
[^(ms, cs)], which could concievably happen with the other arguments.
It also seems that having both a specific hint with the arguments
applied and a general one without doesn't give the speed improvement,
even when the general hint is at higher cost.