write out a first sketch of a Hoare quadruple
we have four things: the precondition, the rely, the guarantee, and the program.

maybe we will need to fold the rely and guarantee into the precondition, if we
need to have them refer to the same variables as the precondition...  probably
would need to do this in the same way as [done] and [crash] are passed around.