slight tweaks to concurrent Hoare tuple
get rid of crash condition; it should just arise as a special kind of
guarantee (or the rely of a special crash-observer thread).

pass rely and guarantee to the precondition, much in the way that
we handled the [done] and [crash] predicates in the original CHL.
this should allow us to use the same set of existential variables
that span pre/rely/guarantee.