fix up [ccorr2] even more, and start stating/proving general theorem about [ccorr2 Write]
need to strengthen the inductive hypothesis when doing induction over the
[star cstep_any] that other threads might have taken before our thread runs.

the strengthening is probably that Write's precondition still holds.

generally, this feels like the right place to require stability.  in this
particular case, our R/G are probably already enough to ensure stability
and thus prove this hypothetical strengthened induction hypothesis.