some more steps towards automating [ccorr2] proofs
prove a [pimpl_cok] theorem, analogue of [pimpl_ok2].

define a [write2] that writes two blocks.

write a spec [write2_cok] and start proving it using [pimpl_cok] using
the recently-proved [write_cok] as a building block.

annotate a bunch of places where we seem to have some issues remaining.

aside: i wonder if we should get rid of thread IDs, and simply have the
notion of a concurrent execution of threads (p1 || p2)..  the thread IDs
seem slightly artificial, and not seemingly needed for anything.