State a parallel composition theorem
This is what we think will allow us to prove that running a set of
programs with Hoare and rely/guarantee tuples with cooperating rely and
guarantee conditions results in an outcome that satisfies the
postconditions.

Theorem parallel_composition itself is unlikely to be the final say, but
proving it is a good exercise and valuable to test our notion of
execution.