Strengthen read_bytes spec to include progress
In non-error cases, spec also promises to read at least one byte.
Unfortunately this is done in a strange way with a postcondition that is
an implication; a separate "valid parameters" spec could move these
extra hypotheses to the precondition.

Somewhat remarkably the automation completely handles this obligation
using omega, so the proof is the same.