Automate ccorr2_{,stable_}step proofs
Unfortunately the pres_step statement no longer quite works, since it
requires that the memory is equal to m', which is clearly not stable
under rely. Looks like we need a weaker pres_step that only requires
rely m mthis, but this might be problematic for the main compose proof.