Track events in cexec
Event list itself doesn't attribute acquires/releases to a thread, but
induction over execution will tell you which program's they are by
virtue of which constructor was used. Note that the new_events added by
rstep is always empty or a singleton acq or rel, just looking at all the
constructors of rstep.