Add fine-grained lock enforcement to semantics
Semantics enforces a StateR across yield points and a LockR at every
write, as well as corresponding StateI and LockI. The LockR/LockI are
intended solely for enforcing that every program respects locks for
every write, so that async reads can trust that the locking protocol is
followed.

To make writes (Assgn/Write) work with Commit, it is now required that
writes are _preceded_ by a Commit that respects the lock discipline
(that is, modifies the ghost copy when it is allowed to do so). All the
code has been updated to do so. Commit will eventually be renamed so
this pattern doesn't look odd.

Writeback has some issue with the intermediate Write not respecting the
cache LockI. This is due to the cache enforcing the fairly sophisticated
policy that the virtual disk reflect the disk + ghost cache, which is
probably not necessary. It should work out in this case regardless,
although might require an additional Commit.