Start changing IO safety to use conflict detection
The major accomplishment so far is to implement Read as a combination of
StartRead and FinishRead, which create the read record and retrieve the
value respectively. AsyncRead should be implementable as BeginRead;;
Yield;; FinishRead. This simplifies the formulation of the semantics,
which will improve further when we unify Yield and AcquireLock into a
WaitForPred primitive.