replace [R_stutter] with [R_trans]
instead of requiring that a relation always admit empty steps, say
something stronger: a relation must be closed under transitivity
(including zero steps of the relation).

this seems important for specs of functions that might or might not yield.
for instance, consider this situation with [AcquireLock]:

the old [AcquireLock] would always appear to yield, even if it didn't
need to.  this means that it always resets the starting ghost state s0.

when we are implementing the new non-primitive [AcquireLock], it actually
does not call [Yield] in some cases, which means it doesn't reset the
starting ghost state s0.

as a result, the spec for the non-primitive [AcquireLock] must consider
two cases: either it yielded internally (and reset s0), or it didn't
yield internally (and thus didn't reset s0).  this is messy.

the same situation will likely arise with any function that might (but
doesn't necessarily) yield, such as cache_read or cache_write, or any
function on top of them.

requiring [R_trans] means we can write specs that make it seem like
the yield happened, but infact it would be perfectly fine if there was
no yield too.