Progress toward hiding readers from R
Attempts to formalize notion that R doesn't read the disk's readers,
running into some trickiness coming up with a property that is general
enough and true of the cache locking discipline. The complexity should
mostly be hidden above the cache, where readers are irrelevant and
indeed can be hidden.

The current approach is to just have precise specs and simplify them at
a higher level, which could be sufficient depending on what we do.