revise crash conditon for commit_ok
If crash happens after DLog.extend succeeded, but before returning,
we are left in a wired state where the active transaction has been
applied, but the in-memory state for current transaction (MSCur ms)
hasn't been reset to map0.  The state is different from that in the post
condition (NoTxn m2), but (ActiveTxn m2 m2).  Fortunately, all memory
state will be reset after crash, so it will go back to (NoTxn m2) upon
recovery.